requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blobfile==2.0.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
discord-py-interactions==4.4.1

docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
html==1.13
huggingface-hub==0.11.1
idna==3.4
importlib==1.0.4
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openai==0.27.4
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.17
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tokenizers==0.13.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
yarg==0.1.9
yarl==1.8.2